put settings in different files
