+++ src/imgui/big_picture.h
@@ -4,8 +4,6 @@
 #pragma once
 
 #include <filesystem>
-#include <map>
-#include <vector>
 #include <SDL3_image/SDL_image.h>
 
 namespace BigPictureMode {
@@ -18,47 +16,8 @@ struct Game {
 };
 
 void Launch();
-void SetGameIcons(bool ForSettings = false);
+void SetGameIcons();
 void GetGameInfo();
 std::filesystem::path UpdateChecker(const std::string sceItem, std::filesystem::path game_folder);
-
-// Settings
-enum class SettingsCategory {
-    Profiles,
-    General,
-    Experimental,
-};
-
-struct Textures {
-    SDL_Texture* general;
-};
-
-void LoadDataToTexture(std::string resourcePath, SDL_Texture*& texture);
-void DrawSettings();
-void SaveSettings(std::string profile);
-void LoadSettings(std::string profile);
-void LoadCategory(SettingsCategory);
-
-void AddCategory(std::string name, SDL_Texture* texture, SettingsCategory category);
-void AddSettingBool(std::string name, bool& value);
-void AddSettingSliderInt(std::string name, int& value, int min, int max);
-void AddSettingCombo(std::string name, int& value);
-int GetComboIndex(std::string selection, std::vector<std::string> options);
-
-//////////////////// option maps for comboboxes
-
-const std::vector<std::string> optionsLogType = {"sync", "async"};
-
-const std::map<std::string, std::vector<std::string>> optionsMap = {
-    {"Log Type", optionsLogType},
-};
-
-//////////////////// Settings struct
-// Note: use int instead of std::string for all combo settings as needed by ImGui
-struct CurrentSettings {
-    bool logEnabled;
-    int volume;
-    int logType;
-};
 
 } // namespace BigPictureMode
